Adds support for variable sized allocas.
Adds support for variable sized allocas. For a variable alloca, code is inserted to first check if the current stacklet has enough space. If so, space is allocated by simply decrementing the stack pointer. Otherwise a runtime routine (__morestack_allocate_stack_space in libgcc) is called which allocates the required memory from the heap. The runtime also does some bookkeeping to ensure we don't leak memory.