Stephen vs. the register allocator, round N+1.
If I emit instructions with vregs for the swapstack code, they will
occasionally be spilled. This is bad, as for that code segment the
stack doesn't exist.

Also, LLVM's support for the reg-reg XCHG instruction is a little
dodgy, and it's unclear how to express the fact that both operands
are modified.

For now, quick hack involving only MOVs and hardcoded registers, to
be improved later.